part1/sensor_fft_consumer/requirements.txt
kafka-python==2.0.2
scipy==1.10.1